+++ lewp/src/config/mod.rs
@@ -1,5 +1,7 @@
 //! Configuration data structures used by [lewp](crate).
 
+use {crate::fh::FileHierarchy, std::sync::Arc};
+
 /// Various config values for adjusting the modules behavior.
 pub struct ModuleConfig {
     /// If true, the module gets wrapped by a `<div>` tag.
@@ -25,17 +27,22 @@ impl Default for ModuleConfig {
 pub struct PageConfig {
     /// Creates a default viewport tag and appends it on rendering.
     pub viewport_tag: bool,
+    /// The file hierarchy where the module is working on.
+    pub fh: Option<Arc<FileHierarchy>>,
 }
 
 impl PageConfig {
     /// Creates a new instance with default values.
-    pub fn new() -> Self {
-        Self { viewport_tag: true }
+    pub fn new(fh: Option<Arc<FileHierarchy>>) -> Self {
+        Self {
+            viewport_tag: true,
+            fh,
+        }
     }
 }
 
 impl Default for PageConfig {
     fn default() -> Self {
-        Self::new()
+        Self::new(None)
     }
 }